backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
aiohttp==3.8.5
requests==2.31.0
finnhub-python==2.4.19 